tap: fix build of tap_bpf_program
The tap_bpf_program.c is not built as part of normal DPDK
EAL environment. It is intended to be built standalone
and does not use rte_common.h.

This reverts the related change from
commit ef5baf3486e0 ("replace packed attributes")

Note: this patch will cause expected warnings from checkpatch
because the code involved is not used directly in DPDK environment.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>