doc: announce changes to event device structures
The structures ``rte_event_dev_info``, ``rte_event_fp_ops`` will be
modified to add new elements to support link profiles.
A new field ``max_profiles_per_port`` will be added to
``rte_event_dev_info`` and ``switch_profile`` will be added to
``rte_event_fp_ops``.

A profile is a unique identifier for a set of event queues linked to
an event port. The unique identifier spans from 0 to the value
advertised in ``rte_event_dev_info.max_profiles_per_port`` - 1.

Two new experimental APIs will be added, one to associate a set of
event queues with a profile which can be linked to an event port and
another to switch the profile which would affect the next dequeue call.

Signed-off-by: Pavan Nikhilesh <pbhagavatula@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>