net/netvsc: increase VSP response timeout to 60 seconds
The current timeout is set to 5 seconds. In Azure, tests show that it may
take up to 15 seconds for VSP to respond on busy nodes. The VSP schedules
unbounded work to process VSC resquest, there is no upper limit on how long
it takes to send response back to VSC.

In the NETVSC kernel mode driver, it waits forever for VSP response. While in
DPDK we can't wait forever, setting the timeout to 60 seconds.

Cc: stable@dpdk.org
Signed-off-by: Long Li <longli@microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>