build/x86: remove conditional checks for AVX2 support
In a number of libraries and drivers we have multiple levels of checks
for enabling AVX2 support. In these cases:

* we first check for AVX2 support in the build-time ISA, i.e. through
  the instruction-set/-march flag. If present, we add source file to the
  list of sources.
* if not enabled at the minimum instruction-set level, i.e. a
  default/generic build, we then check for compiler support for AVX2
  and, if available, did a separate build of the AVX2 file using an
  additional flag to enable the instruction set.

While this works, and was necessary in older releases, we no longer need
this level of complexity, as all supported DPDK compilers have support
for AVX2. This makes the second check unnecessary.

However, when we look at the effect of the second option above vs the
first, the only real difference is that in the second case, we add an
additional "-mavx2" to the build flags. This flag simply makes the
instruction set available for use, so should be harmless in the case
where the "march" flag already has that instruction set available.
Therefore, we can remove the check for the first case also, and always
use the fallback case with the extra flag.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>