build: deprecate enable_kmods option
With the removal of the kni kernel driver, there are no longer any
Linux kernel modules in our repository, leaving only modules for FreeBSD
present. Since:

* BSD has no issues with out-of-tree modules and
* There are no in-tree equivalents for those modules in BSD

there is no point in building for BSD without those modules.

Therefore, we can remove the enable_kmods option, always building the
kmods for BSD.

We can also remove the infrastructure for Linux kmods too, since use of
out-of-tree modules for Linux is not something the DPDK project wants to
pursue in future.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>