net/iavf: add devargs to enable vf auto-reset
Originally, the iavf PMD does not perform special actions when it
receives a PF-to-VF reset event, resulting in vf being offline and
unavailable.

This patch enables vf auto-reset by setting 'watchdog_period' devargs
to true. The iavf PMD will perform an automatic reset to bring the vf
back online when it receives a PF-to-VF event.

v2: using event handler to handle reset

Signed-off-by: Shiyang He <shiyangx.he@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>