examples/l3fwd: make eventdev an optional dependency
With l3fwd being a very commonly used example app, and built as part of
our CI tests, we need to ensure it's buildable with just about all
supported DPDK configurations.  To enable l3fwd application to be built
when the eventdev library is disabled, we need to compile in the
eventdev support conditionally.  Thankfully, the eventdev support is
pretty self-contained, with only the main.c file having more than a
couple of ifdefs.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Acked-by: Morten Brørup <mb@smartsharesystems.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>