net/iavf: add support for runtime queue reconfiguration
Unlike the i40e driver, the iavf driver does not advertise support for
runtime reconfiguration, meaning that application using the same
hardware may get different behaviour when using a PF vs a VF. On testing
with a 40G NIC, the only blocker to reconfiguring an RX queue on the fly
is the fact that this support is not advertised by the driver.

Add support for runtime reconfig by reporting it in the device info
flags.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>